requirements.txt
tensorflow==2.16.2
# tensorflow-metal
tensorflow-probability==0.19.0
tensorflow-addons
tf_keras
numpy 
matplotlib
deeplake
pandas
opencv-python
torch 
onnx
tf2onnx
onnx-tf